dns/opendnssec: adding EOL message to opendnssec-1.4.14
The upstream has declared this soon to be EOL. The port is now updated accordingly.

End-of-life announcement: https://www.opendnssec.org/2019/10/

OpenDNSSEC 2.1 was released in February 2017, and in the past
two-and-half year it has proven itself to be stable and viable
upgrade of 1.4, and has additional features and improvements.

Therefore we announce end-of-life of OpenDNSSEC 1.4. One of the steps
towards future releases with better experience and shorter cycles.

Starting today, October 8, 2019, in accordance with our policies,
we will only provide essential fixes and support until 9 October 2020,
after which support will no longer be available. We feel confident
that existing installations can upgrade without much hassle and offer
support to our customers in doing so.

OpenDNSSEC 2.1.4 [1] serves as the replacement for the 1.4 LTS.
The current version is 2.1.6. There is a migration step necessary,
for which you can find a good breakdown at the migration page. [2]

[1] https://svnweb.freebsd.org/ports/head/dns/opendnssec2
[2] https://www.opendnssec.org/migration-from-1-4-to-2-1/

PR:		244610
Submitted by:	Jaap Akkerhuis <jaap@NLnetLabs.nl> (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@528036 35697150-7ecd-e111-bb59-0022644237b5