www/py-django111: Mark as DEPRECATED
Django 1.11 has reached its End-of-Life so mark it as DEPRECATED.  Don't set
an EXPIRATION_DATE yet as various consumers need to be switched to
Django 2.2 beforehand, which will happen in the next weeks/months.

When everything is complete there should be no py-dj22-* ports left and the
Ports tree has a few Python 2.7 packages less because Django 2.2 only
supports Python 3.5 or later.

Adjust the description of Django 1.11 of the www/py-django meta-port
accordingly.  Also fix a small typo in the Django 2.2 description while I'm
here.

PR:		245309
Reviewed by:	koobs, wen (via private e-mail)
Differential Revision:	https://reviews.freebsd.org/D24221