audio/musescore prep-work for CMake 3.17
Prepare for building with CMake 3.17, which suffers from Musescore
upcasing the BUILD_TYPE halfway through. Bug report to CMake is
	https://gitlab.kitware.com/cmake/cmake/-/issues/20539
(but I think it's a Musescore problem).

No PORTREVISION bump because nothing changes in current builds,
and this affects only builds that happen after CMake 3.17 comes out.
(and even then the package remains the same)