[NEW] net-mgmt/py-aggregate6:Compress an unsorted list of IPv4 and IPv6 prefixes
Takes a list of IPv6 prefixes in conventional format on stdin, and performs two
optimisations to attempt to reduce the length of the prefix list. The first
optimisation is to remove any supplied prefixes which are superfluous because
they are already included in another supplied prefix. The second optimisation
identifies adjacent prefixes that can be combined under a single, shorter-length
prefix. The above optimalisation steps are often useful in context of
compressing firewall rules or BGP prefix-list filters.

WWW: https://github.com/job/aggregate6