gstreamer1-plugins-v4l2: fix for (some) webcams
The gstreamer v4l2 plugin probes the DMA capabilities of the webcam
device via ioctl VIDIOC_EXPBUF, expecting ENOTTY in case that's not
implemented (and will act accordingly). On our cuse, we get EINVAL
for that case (which, from my reading of ioctl(2) woulndn't be correct
here), which gstreamer neither expects nor handles and thus cannot
use this device as a video source. As we cannot fix the kernel on short
notice (cuse is part of the base system), the workaround is to handle
EINVAL like ENOTTY in this specific place. (Tested with an old
Logitech camera).

PR:		236673
Approved by:	maintainer-timeout (multimedia)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@534176 35697150-7ecd-e111-bb59-0022644237b5