textproc/ripgrep: do install completion file for fish
To solve bug 244355, we stopped installing the completion file for fish,
because fish installed it itself.  Now fish has stopped doing that, and they
request that packages like ripgrep install it themselves.  But we are to
install it into a different directory than we used to.  This change puts it
into the new location.

PR:		246290
Approved by:	petteri.valkonen@iki.fi (maintainer)