audio/ardour5: add BUILD_DEPENDS as:devel/binutils
/usr/bin/as will soon be removed from the FreeBSD base system, for
FreeBSD 13.0.

Depend on as (not ${LOCALBASE}/bin/as) so that this will introduce no
change for existing 11.x and 12.x, which will continue to use the base
system GNU as 2.17.50.

This likely also addresses the reason this port is marked BROKEN_aarch64;
that can be addressed in a subsequent sweep.

PR:		205250
Approved by:	portmgr (blanket)
Sponsored by:	The FreeBSD Foundation