Update 3.32.0 --> 3.32.2
Fix two long-standing bugs that allow malicious SQL statements to crash
the process that is running SQLite. These bugs were announced by a
third-party approximately 24 hours after the 3.32.0 release but are not
specific to the 3.32.0 release.

Fix a long-standing bug in the byte-code engine that can cause a COMMIT
command report as success when in fact it failed to commit. Ticket
810dc8038872e212.

Approved by:	portmgr (blanket: run-time bugfix)
MFH:		2020Q2