sysutils/pwol: Add new ports
Add sysutils/pwol, a small tool that can be used to send WoL packages to
wake up hosts or groups of hosts.  It can also act as a WoL gateway when you
need to forward such packages but don't have a router that supports it.

PR:		247251
Submitted by:	Peter Eriksson


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@538926 35697150-7ecd-e111-bb59-0022644237b5