add new port devel/libffi321
Add a new port, devel/libffi321, which is a stripped down version of libffi
3.2.1.  This version only brings in the shared library from libffi 3.2.1
(the libffi 3.3.0 update bumped the version number of the shlib) for binary
consumers that can't easily be recompiled and updated against the new
version of libffi.
This version is not meant to be linked against, and you should not depend on
it to bring libffi in, unless you explicitly need the binary compat.
Libraries are installed in ${LOCALBASE}/lib/compat.

PR:		247028 (for tracking)
Requested by:	arrowd
Tested by:	arrowd


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@541359 35697150-7ecd-e111-bb59-0022644237b5