devel/xtensa-esp32-elf: Update to fix PSRAM cache issue
This version of the toolchain includes fixes for the PSRAM cache
issue described here:

    https://github.com/espressif/esp-idf/issues/2892

Limit builds to tier 1 platforms (amd64 and i386) until we can
switch to a version that builds with gcc9.

Turn on CT_FORBID_DOWNLOAD and remove dependency on git; we never
want to use it to download anything.

Make some incremental improvements to the port Makefile and pkg-plist.