Make qt5-core and qt5-webengine depend explicitly on double-conversion.
In a dirty build-environment, double-conversion was being picked up
automatically (as "system double-conversion"). We could bung in
a configure flag to ignore it, so that Qt would always build its
own copy. Or we could depend explicitly on it, dropping the build
of the bundled copy.

Do the latter, because:
- it's the same code, but now extracted from V8 JS into a separate
  library by upstream (that is, by Chromium / Google),
- the dependency package is only 168kB.
- this reduces code-bundling and improves reuse when other things
  also use double-conversion.

(This is the opposite of what I originally intended; using the flag
is harder to implement, and since the dependency is small ..)

I've taken the patch from Andy Mender and applied it in two cases.
Perhaps adding it to qt5-core would have been enough, duplication
here can't hurt.

PR:		245946 248020
Submitted by:	Andy Mender
Reported by:	Andy Mender, yuri