devel/valgrind-devel: finish Makefile reordering, no functional change
Mea culpa, the rest of this re-ordering is due to having missed this line in
a bunch of portlint noise about the patches that will soon be disappearing:

WARN: Makefile: "USES" has to appear earlier.

Move conflicts/bb/options definitions further down, as is also done in the
conflated diff of D25452.

Approved by:	zeising (maintainer, implicit)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543271 35697150-7ecd-e111-bb59-0022644237b5