New port: devel/rinutils
This is a set of C headers containing macros and static functions that are
expected to work on Unix-like systems and MS Windows that have been extracted
from Shlomi Fish´s projects.

They are quite random and task specific and may not work for you.

WWW: https://github.com/shlomif/rinutils

This is a requirement for an upcoming dependency in games/kpat


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543456 35697150-7ecd-e111-bb59-0022644237b5