Enable -shared support for embedded RISC-V targets
Shared library support is incomplete for the RISC-V embedded elf
targets, and was disabled some time ago [1]. However, it appears that it
is functional enough to support basic usage, and is required to
successfully build u-boot for RISC-V. This toolchain doesn't have
widespread usage, so any risk of re-enabling this support is outweighed
by the benefit of being able to build important ports like u-boot.
Override this default.

[1] https://sourceware.org/git/?p=binutils-gdb.git;a=commit;h=be1dce26256aa9f6c7742e81c8c3db9eacff079d

Reviewed by:	philip
Approved by:	philip
Differential Revision:	https://reviews.freebsd.org/D25735


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543868 35697150-7ecd-e111-bb59-0022644237b5