New port: x11-fonts/libertinus
The Libertinus fonts project began as a fork of the Linux Libertine and Linux
Biolinum fonts. The original impetus was to add an OpenType math companion to
the Libertine font family. Over time it grew into to a full-fledged fork
addressing many of the bugs in the Libertine fonts.

WWW: https://github.com/alerque/libertinus