devel/libffi: Fix build on i386 without objdump
FreeBSD 13 does not have any objdump in the default installation, which
causes issues with libffi on i386 with the recent update to llvm 11.
Add a patch originally from upstream, to use readelf instead of objdump.

The patch is for configure.ac originally.  I've applied it and rerun
autoreconf and committed the result, instead of adding a dependency on
autoreconf.

Fix devel/libffi321 in the same way.  While here, fix the fuzz on a few
other patches, which made it much easier to sort out the configure patch.

Thanks to dim@ for helping me figure out the issue.

MFH:		2020Q3