lang/ocaml: Fix build with llvm 11
Fix the build of lang/ocaml with llvm 11.
There are two issues, first off, te issue with -fno-common, which is default
with llvm 11.  This is fixed by pulling in a modified patch from upstream,
as well as local patches.
Secondly, there's a problem with how asm sections are handled, patch this
locally.  Thanks to dim@ for help with this.

MFH:		2020Q3 (implicit, -fno-common fixes, ok by joenum)