grapics/drm-legacy-kmod: Deprecate
Deprecate graphics/drm-legacy-kmod.  Update the drivers to include a
deprecation message when the drivers are loaded.

drm-legacy-kmod is holding back changes in the FreeBSD VM subsystem, and it
requires substantial changes to be updated to work with the VM changes.
See the e-mail to x11@FreeBSD.org for more information.

MFH:		2020Q3 (implicit, drm-drivers blanket)