Extend expiration of ports broken with -fno-common
Extend the expiration of unmaintained leaf ports that are broken with
-fno-common and previously had a short expiration date.  The new date is set
to 2021-04-01, meaning the ports will be in the 2021Q2 branch, giving people
until 2021-06-30 (10 months from now) to fix issues.
Improve the wording of the deprecation message, mentioning both llvm 11 and
-fno-common.
Bump portrevision to force reinstall.  This way the warning message about
deprecation will be displayed to users installing or reinstalling the
software.

MFH:		2020Q3 (maybe)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@547044 35697150-7ecd-e111-bb59-0022644237b5