Disable python. This ancient version of gdb wants python2.
Also, while I'm here drop maintainership. I've not used this port in 4 years.
Mark it as DEPRECATED and expire at the end of the year since I don't think it's
worth updating. If someone else wants to do it, be my guest. If there's a better
expiration date, then feel free to update it.

Also, appease portlint while I'm here...