x11/libxcb: Re-add runtime dependency on pthread-stubs
Re-add the runtime dependency on pthread-stubs, that was dropped in the
update to 1.14.  Things linking against libxcb needs it as well, and since
it was pulled in from here, keep doing that.

In the future, the dependency on pthread-stubs mabye should be moved to the
ports that actually need it, but keep the status quo for now.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@548493 35697150-7ecd-e111-bb59-0022644237b5