graphics/piglit: Update to latest upstream
Update to upstream. Patches comes from a MR opened upstream.
Remove opencl dependancy, I haven't try it yet and will be dealt with
in a futur update.
Remove wayland option, wayland is always needed now. piglit is a test
suite and we want to test everything, it's also not intented for users
to use but only developpers.

Reviewed by:	x11 (zeising@)
Differential Revision:	https://reviews.freebsd.org/D26396


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@548592 35697150-7ecd-e111-bb59-0022644237b5