Simplify www/qt5-WebEngine dependencies.
WebEngine contains a Qt Designer plugin -- so you can put webengine
widgets into forms in designer -- so it needs designer to build that
plugin against. However, you don't need designer to use webengine
parts -- as a web-browser, it doesn't use the designer application
at all. So shuffle the designer dependency to _build; this means
that installing a web browser based on webengine no longer pulls
in the designer application.

Of course, you'll need to install designer, in order to use the
webengine designer plugin, but that's for specific devel-work.

PORTREVISION bumped because the deps change.

Reported by:	FuryBSD


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@548883 35697150-7ecd-e111-bb59-0022644237b5