Always prevent building valadoc.
The build system check for valadoc, if found in the environement then
it will be used to generate the devhelp documentation. otherwise no.
This behaviour is not optionable, so disable entirely the generation
with a patch for now.

note that in a clean environement it does not happen as valadoc is not
present.

PR:		249423
Reported by:	mandree


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@548905 35697150-7ecd-e111-bb59-0022644237b5