Fix build of www/nginx-full with -fno-common
This fix should also apply to www/nginx built with option HTTP_PUSH, but
the compilation fails with:

    error: no member named 'sockaddr' in 'struct redisContext'

This seems to be an unrelated issue and I have not tried to diagnose it,
yet.

The PORTREVISION is only bumped for www/nginx-full, which built for me with
this patch applied.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549093 35697150-7ecd-e111-bb59-0022644237b5