Fix build with -fno-common
Since files that are patched in legacy patch files are affected, all
patches are regenerated with make makepatch.

In order to have makepatch operate on only the patched sources without the
result of RE_INPLACE being integrated into the patch files, move the
in-place editing of sources from post-patch to pre-build.