Fix build with -fno-common and correct software bugs
There were software bugs in 3 source files, most due to wrong use of sizeof
(e.g. using the size of a pointer instead of the data, or subtracting a
value from the argument of sizeof, not the resulting size value.

A test for a short write assigned the length written to a variable and
performed a comparison with that variable in a single expression and with
no defined order of the these two operations resulting in either a
comparison with 0 (the value before assignement) or with the just assigned
value (tautological comparison). Either case did not catch a short write.

I have not checked the quality of the code nay further than these issues
that caused compiler warnings, but given the severity and fundamental lack
of understanding shown by these examples, I'm not convinced that this
program can be trusted to work correctly.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549729 35697150-7ecd-e111-bb59-0022644237b5