devel/pystring: create port
Pystring is a collection of C++ functions which match the interface and
behavior of python's string class methods using std::string. Implemented in
C++, it does not require or make use of a python interpreter. It provides
convenience and familiarity for common string operations not included in the
standard C++ library. It's also useful in environments where both C++ and
python are used.

WWW: https://github.com/imageworks/pystring


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549776 35697150-7ecd-e111-bb59-0022644237b5