Fix and update broken port
This is ancient software that has been made available and somewhat updated
on GitHub. The previous port was broken and with this update it should at
least have a chance to compile.

Testing on amd64 gave reasonable results, much to my surprise, since I had
doubt that the code is 64 bit safe. The programming style is strange, with
C files being concatenated by #including them into the main program. This
is definitely a unique piece of software ... (might serve as an dissuasive
example).

Since this program has to be run with root privileges and may well have
security weaknesses, I'm not convinced that we should even provide a port
and package.

But I'm committing this update anyway, leaving to somebody else the decision
to remove it from ports, if considered too much of an annoyance or risk.