devel/nlohmann-json: Install include/nlohmann/json_fwd.hpp
This header file is required by future versions of sysutils/nix.

In order to get that header installed, the JSON_MultipleHeaders option had
to be enable. This option results in the build system installing a lot of
other header files as well. As there is no need to ship those other headers
at the moment, let's delete them in post-install.

PR:		249895
Approved by:	yuri@ (maintainer)