bsd.port.mk: Handle incorrect binary aliases gracefully
Currently, if the src element of a BINARY_ALIAS pair is missing,
the ports framework would just print out the help message of install(1)
and exit with a failure. This might be hard to debug for novice ports
users.

Let's introduce a check, which would allow us to print
that the src component is missing.

This situation happens more often than it seems to. For example,
BINARY_ALIAS tend to blow up when used for test dependencies,
because binary aliases are created before test dependencies
are installed. This leads to a lot of confusion for my experience.

Approved by:	portmgr (mat)
Differential Revision:	https://reviews.freebsd.org/D26551


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@550426 35697150-7ecd-e111-bb59-0022644237b5