Try to fix build on recent FreeBSD versions where clang version > llvm default ports version. Fix this by depending on a higher version of llvm if such a version exixts, and otherwise use a clang version compatible with the latest llvm.
Using llvm was previously dependant on clang being used, and this did not quite
work for all cases due to some weird order of variables in the Makefile. The
result was that the option set by the used was not really honored. The
portrevision is bumped to reflect that the setting of LLVM might not have
worked before this portrevison.

PR:	244403


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@550749 35697150-7ecd-e111-bb59-0022644237b5