New port: New sysutils/choria
Choria is a framework for building Control Planes, Orchestration Systems and
Programmable Infrastructure.

This is a daemon and related tools written in Go that hosts services,
autonomous agents and generally provide a secure hosting environment for
callable logic that you can interact with from code.

WWW: https://choria.io/

With hat:	puppet


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@551252 35697150-7ecd-e111-bb59-0022644237b5